+++ cv_pipeliner/inference_models/detection/core.py
@@ -7,18 +7,23 @@ from cv_pipeliner.core.inference_model import ModelSpec, InferenceModel
 
 Bbox = Tuple[int, int, int, int]  # (xmin, ymin, xmax, ymax)
 Score = float
+Class = str
 
 Bboxes = List[Bbox]
 Scores = List[Score]
+Classes = List[Class]
 
 DetectionInput = List[np.ndarray]
 DetectionOutput = Tuple[
     List[Bboxes],
-    List[Scores]
+    List[Scores],
+    List[Classes],  # Optional exit
+    List[Scores]  # Optional exit
 ]
 
 
 class DetectionModelSpec(ModelSpec):
+    class_names: List[str] = None  # optional
 
     @abc.abstractproperty
     def inference_model_cls(self) -> Type['DetectionModel']:
@@ -32,8 +37,12 @@ class DetectionModel(InferenceModel):
         super().__init__(model_spec)
 
     @abc.abstractmethod
-    def predict(self, input: DetectionInput,
-                score_threshold: float) -> DetectionOutput:
+    def predict(
+        self,
+        input: DetectionInput,
+        score_threshold: float,
+        classification_top_n: int = None
+    ) -> DetectionOutput:
         pass
 
     @abc.abstractmethod