+++ changelog.md
@@ -1,9 +1,17 @@
+# 0.5.0 (2021-03-11)
+
+- Detector now can be pipeline itself
+- Add complex pipeline: TextDetection
+
 # 0.4.5 (2021-01-18)
+
 - Package `object_detection` is now not required
 
 # 0.4.4 (2021-01-18)
+
 - Annotation in dataset browser is made better: top-n is now showed in annotation list.
-- Backups in annotation mode are now made every hour. 
+- Backups in annotation mode are now made every hour.
 
 # 0.4.3 (2021-01-13)
+
 - Added quick annotation mode in the main app for model quality research